gebruikers
Status updaten voor gebruikers ga ik niet doen. De routes heb ik weggehaald die daarvoor waren bestemd.
