+++ resources/views/profile.blade.php
@@ -48,8 +48,8 @@
                         </div>
                     @endif
                     You are logged in!
-                    <br><a href="/afspraken/mijn">Mijn afspraken</a>
-                    <br><a href="/opentimeslots/mijn">Mijn open timeslots</a>
+                    <br><a href="/appointments">Mijn afspraken</a>
+                    <br><a href="/timeslots">Mijn open timeslots</a>
                 </div>
             </div>
         </div>
