+++ resources/views/users.blade.php
@@ -1,0 +1,12 @@
+@extends('layouts.app')
+
+@section('content')
+    <h1>Alle gebruikers</h1>
+    <p>
+        <ul>
+            @foreach($users as $user)
+                <div>{{ $user->name }}</div>
+            @endforeach
+        </ul>
+    </p>
+@endsection
