+++ routes/web.php
@@ -21,13 +21,6 @@ Route::get('/profile', 'PagesController@profile');
 
 // Routes met de afsprakencontroller
 Route::resource('/appointments', 'AppointmentsController');
-//Route::get('/appointments', 'AppointmentsController@index');
-//Route::get('/appointments/create', 'AppointmentsController@create');
-//Route::get('/appointments/{appointment}', 'AppointmentsController@show');
-//Route::post('/appointments', 'AppointmentsController@store');
-//Route::get('/appointments/{appointment}/edit', 'AppointmentsController@edit');
-//Route::patch('/appointments/{appointment}', 'AppointmentsController@update');
-//Route::delete('/appointments/{appointment}', 'AppointmentsController@destory');
 
 // Routes met de timeslotscontroller
 Route::get('/timeslots', 'TimeslotsController@index');
@@ -37,14 +30,11 @@ Route::post('/timeslots/store', 'TimeslotsController@store');
 // Routes met betrekking tot authorisatie en authenticatie
 Auth::routes();
 Route::group(['middleware' => ['moderator']], function() {
-    Route::get('/moderators', 'ModeratorsController@index')->name('moderator');
-    Route::patch('/moderators', 'ModeratorsController@promote')->name('moderator');
-    Route::patch('/moderators', 'ModeratorsController@religate')->name('moderator');
+    Route::get('/users', 'UsersController@index')->name('moderator');
 });
 
 // Met betrekking op gebruikers
 Route::get('/users', 'UsersController@index');
 Route::get('/users/{user}/edit', 'UsersController@edit');
 Route::patch('/users', 'UsersController@update');
-Route::patch('/users/{user}', 'UsersController@update');
 Route::delete('/users/{user}', 'UsersController@destroy');
